requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
